beginning fixing tests
