+++ test/browser.js
@@ -30,7 +30,6 @@ test.onFailure(window.browserTestDriver_fail);
 
 test('Browser tests', t => {
   require('./node/index.js');
-  require('./browser/index.js');
-
+  // require('./browser/index.js');
   t.end();
 });
