+++ test/helpers/comparison-utils.js
@@ -151,7 +151,7 @@ export function cmpDataset(t, expectedDataset, actualDataset, opt = {}) {
   // test everything except auto generated color
   Object.keys(actualDataset).forEach(key => {
     if (key === 'fields') {
-      cmpFields(t,  expectedDataset.fields, actualDataset.fields, expectedDataset.id);
+      cmpFields(t, expectedDataset.fields, actualDataset.fields, expectedDataset.id);
     } else if (key === 'gpuFilter') {
       // test gpuFilter props
       cmpGpuFilterProp(t, expectedDataset.gpuFilter, actualDataset.gpuFilter);
@@ -254,20 +254,12 @@ export function cmpParsedAppConfigs(t, expectedConfig, actualConfig, {name} = {}
 }
 
 export function cmpFields(t, expected, actual, name) {
-  t.equal(
-    expected.length,
-    actual.length,
-    `dataset.${name} should have same number of fields`
-  );
+  t.equal(expected.length, actual.length, `dataset.${name} should have same number of fields`);
   actual.forEach((actualField, i) => {
-    cmpField(
-      t,
-      expected[i],
-      actualField,
-      `dataset.${name} fields ${actualField.name}`
-    );
+    cmpField(t, expected[i], actualField, `dataset.${name} fields ${actualField.name}`);
   });
 }
+
 export function cmpField(t, expected, actual, name) {
   cmpObjectKeys(t, expected, actual, name);
 
@@ -292,7 +284,7 @@ export function cmpField(t, expected, actual, name) {
         });
       }
     } else if (k === 'valueAccessor') {
-      t.ok(typeof actual[k] === 'function');
+      t.ok(typeof actual[k] === 'function', `${name}.valueAccessor should be a function`);
     } else {
       t.deepEqual(actual[k], expected[k], `${name}.${k} should be the same`);
     }
