+++ test/node/utils/index.js
@@ -20,17 +20,16 @@
 
 import './data-utils-test';
 import './data-processor-test';
+import './kepler-table-test';
 import './filter-utils-test';
-import './gpu-filter-utils-test';
-import './dataset-utils-test';
-import './layer-utils-test';
-import './data-scale-utils-test';
-import './interaction-utils-test';
-import './mapbox-gl-style-editor-test';
-import './notifications-utils-test';
-import './aggregate-utils-test';
-import './color-util-test';
-import './util-test';
-import './export-utils-test';
-import './s2-utils-test';
-import './map-info-utils-test';
+// import './gpu-filter-utils-test';
+// import './layer-utils-test';
+// import './data-scale-utils-test';
+// import './interaction-utils-test';
+// import './mapbox-gl-style-editor-test';
+// import './notifications-utils-test';
+// import './aggregate-utils-test';
+// import './color-util-test';
+// import './util-test';
+// import './export-utils-test';
+// import './s2-utils-test';
